pyproject.toml
[project]
name = "shelf-genius"
version = "0.1.0"
description = "AI agent that give book suggestions based on a image of your book shelf"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.13",
    "langchain-core>=0.3.28",
    "langgraph>=0.2.60",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.3",
    "pydantic>=2.6.1",
    "pillow>=10.2.0",
    "langchain-openai>=0.3.0",
    "requests>=2.32.3",
    "langchain-anthropic>=0.3.1",
    "langchain-groq>=0.2.3",
]

[tool.hatch.version]
path = "src/shelf_genius-genius/__about__.py"

[dependency-groups]
dev = [
    "hatch>=1.14.0",
    "pytest-asyncio>=0.25.1",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
packages = ["src/shelf_genius"]

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default
lint.select = ["E", "F"]
# Allow unused variables when underscore-prefixed
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Python version compatibility
target-version = "py310"
# Line length matches black's default
line-length = 120

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
